+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Improved color variables.
 - Added more `libraries-override` to reduce amount of unused styles on the page.
 - Improved system libraries override and added jQuery UI dialog clean styles.
+- Added custom AJAX indicator for easier theming.
 
 ## [2.0-alpha2]